override storage instead of disable
